Add support for setting the working directory of tasks on Linux
Glibc 2.29 will add posix_spawn_file_actions_addchdir_np. A method with
the same signature also exists on Solaris, and there is currently a
proposal to add a "posix_spawn_file_actions_addchdir" to POSIX, so
proactively re-structure the code to prefer using a
posix_spawn_file_actions_addchdir[_np] style function to set the working
directory.

We fall back to an alternate mechanism where available, or continue to
assert as before.

rdar://46069443